+++ td/telegram/net/NetQueryDispatcher.cpp
@@ -271,7 +271,7 @@ bool NetQueryDispatcher::is_dc_inited(int32 raw_dc_id) {
   return dcs_[raw_dc_id - 1].is_valid_.load(std::memory_order_relaxed);
 }
 int32 NetQueryDispatcher::get_session_count() {
-  return max(G()->shared_config().get_option_integer("session_count"), 1);
+  return max(narrow_cast<int32>(G()->shared_config().get_option_integer("session_count")), 1);
 }
 
 bool NetQueryDispatcher::get_use_pfs() {
